pyproject.toml
[project]
name = "twitter-screenshot-archive"
version = "0.1.0"
description = "Full-text search for a Twitter screenshot archive"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=12.1.1",
    "pillow-heif>=1.3.0",
    "datasketch>=1.6.5",
    "psycopg>=3.3.3",
    "pytesseract>=0.3.13",
    "pyyaml>=6.0.3",
    "tqdm>=4.67.3",
]

[project.optional-dependencies]
mcp = ["mcp>=1.26", "scikit-learn>=1.3", "numpy"]

[project.scripts]
ingest = "twitter_screenshot_archive.core.ingest:main"
web = "twitter_screenshot_archive.web.app:main"
mcp = "twitter_screenshot_archive.mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/twitter_screenshot_archive"]
